Feat: Finish kelbi blacklisted and auth backup

+++ src/commands/mhfrontier/backup-kelbi.js
@@ -2,6 +2,8 @@ const { SlashCommandBuilder } = require('discord.js')
 const { exec } = require('child_process')
 const fs = require('fs')
 const AdmZip = require('adm-zip')
+const pool = require('../../utils/dbConnection')
+const bcrypt = require('bcrypt')
 const backupFileName = `kelbi_${new Date().toISOString().slice(0, 10)}.sql`
 
 module.exports = {
@@ -9,39 +11,83 @@ module.exports = {
     .setName('backup-kelbi')
     .setDescription('Send a backup of the kelbi database'),
   async execute (interaction) {
+    const loginInput = interaction.options.get('login').value
+    const passwordInput = interaction.options.get('password').value
+    const blacklistInput = interaction.options.get('blacklist').value
+
     await interaction.deferReply({ ephemeral: true })
 
     const tableBlackList = ['quests', 'questlist']
-    const tabelasExcluidasArgs = tableBlackList.map(tabela => `-T${tabela}`).join(' ')
+    const tabelasExcluidasArgs = tableBlackList.map(tabela => `-T ${tabela}`).join(' ')
+    const toogleTableBlacklist = blacklistInput ? '' : tabelasExcluidasArgs
+
     const zipFileName = backupFileName.replace('.sql', '') + '.zip'
-    const cmd = `pg_dump -h ${process.env.DB_HOST} -U ${process.env.DB_USER} -d ${process.env.DB_DATABASE} ${tabelasExcluidasArgs} -f ${backupFileName}`
+    const cmd = `pg_dump -h ${process.env.DB_HOST} -U ${process.env.DB_USER} -d ${process.env.DB_DATABASE} ${toogleTableBlacklist} -f ${backupFileName}`
 
     try {
-      exec(cmd, (error) => {
-        if (!error) {
-          const zip = new AdmZip()
-          zip.addLocalFile(backupFileName)
-          zip.writeZip(zipFileName)
+      const authLogin = await userIsDev(loginInput, passwordInput)
 
-          interaction.editReply({
-            content: 'Backup was done successfully!',
-            files: [zipFileName]
-          })
+      // Discord can't take questlist data becase 8MB limite
+      if (blacklistInput) {
+        interaction.editReply('File too big :/')
+        return
+      }
 
-          setTimeout(() => {
-            if (backupFileName) {
-              fs.unlinkSync(backupFileName)
-            }
-            if (zipFileName) {
-              fs.unlinkSync(zipFileName)
-            }
-          }, 20000)
-        } else {
-          interaction.editReply('Fail to get backup!')
-        }
-      })
+      if (authLogin) {
+        exec(cmd, (error) => {
+          if (!error) {
+            const zip = new AdmZip()
+            zip.addLocalFile(backupFileName)
+            zip.writeZip(zipFileName)
+
+            interaction.editReply({
+              content: 'Backup was done successfully!',
+              files: [zipFileName]
+            })
+
+            setTimeout(() => {
+              if (backupFileName) {
+                fs.unlinkSync(backupFileName)
+              }
+              if (zipFileName) {
+                fs.unlinkSync(zipFileName)
+              }
+            }, 20000)
+          } else {
+            interaction.editReply('Fail to get backup!')
+          }
+        })
+      } else {
+        interaction.editReply('Invalid user or not a dev')
+      }
     } catch (error) {
       interaction.editReply('Unexpected error trying to get backup!')
     }
   }
+}
+
+const userIsDev = async (user, password) => {
+  const query = `
+    SELECT
+      username,
+      password,
+      dev
+    FROM
+      users
+    WHERE
+      username = $1;
+  `
+
+  const { rows } = await pool.query(query, [user])
+
+  if (rows.length > 0) {
+    const dbPassword = rows[0].password
+    const dbDev = rows[0].dev
+
+    const loginMatch = await bcrypt.compare(String(password), String(dbPassword))
+
+    if (loginMatch && dbDev) return true
+  }
+
+  return false
 }
